Use shared volume to cache packages
GitHub has a limit of 10GB of cache per repository. Plus caches are
scoped to a key and a branch, so a cache created on branch-A won’t be
accessible by branch-B unless branch-A is the default branch. Since
llvm-target is not the default branch, we cannot use GitHub cache
directly.