Run "Build and test" workflow on pushes to branch llvm-target
Verify that PR merged to llvm-target do not break that branch. It is
required, specifically, for this case:

A developer made a PR, which was verified by CI and marked as green.
The PR waits for some time in "Needs approval" state, meanwhile genx
branch receives new commits. The PR gets approved and merged.

At this moment we need to verify, that previously approved and verified
PR works with the latest llvm genx commit.