minimize token permissions in workflows (#308)
For https://github.com/intel/intel-xpu-backend-for-triton/issues/300<!--
Thank you for contributing to the project. Here is a small check lists
for you to make the review process quicker: -->

<!-- # Coding Style
pre-commit helps for checking code style and auto-formatting. Please run
it locally before submitting the PR. You could follow
[CONTRIBUTING.md](../CONTRIBUTING.md#coding-style-and-precommit) guide
for installing a hook in your local environment. -->

<!-- # Please Add the following tag on title prefix: -->
<!-- [BACKEND] / [FRONTEND] / [OPTIMIZER] / [CI] / [FIX] / [DOC] -->

<!-- # Description of the PR -->
<!-- Please give a brief description on this PR.
1. What issue does this PR solves?
2. What feature does this PR introduces?
3. Is there a working command/example for checking the PR? For example,
what is the command to run the test? -->