Reduce problem size for tutorial `06-fused-attention.py` (#350)
The tutorial `06-fused-attention.py` has a problem size that is too
large and slow, given that currently the `tl.dot` operator is lowered
using a loop containing scalar FMA instructions. Once `tl.dot` can be
lowered to use DPAS instructions we should be able to change the problem
size back to the original. Meanwhile this PR provides a way to test this
tutorial.

---------

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>
Co-authored-by: Whitney Tsang <whitney.tsang@intel.com>