Add intel backend (triton gpu to llvm conversion) (#541)
This is the first step to move our flow to the backend. Since the first
two attempts failed miserably this does the easiest thing possible -
move the existing pipeline to the backend resolving all the name
collisions.

Since we use nvgpu&friends in our pipeline, I've created a copy of those
in the backend. This proved to be necessary when merging upstream
changes. We will keep it there and gradually clean up.

I kept namespace changes to a minimum (some of the resolution is done by
simple renaming) partly because of potential ambiguity and partly
because most of it will go away with the next steps (mostly related to
utilities which will be split into generic ones in lib/conversion and
backend-specific).