[E2E] Cache for pip cache, torch vision, text, audio (#563)
A single E2E workflow currently launches 30 jobs in matrix. Adding cache
for pip cache, torch vision, text, audio. The key for pip cache is per
commit id, so the first finished job will populate the cache, others
will use it. This is to minimize network traffic, since there is no need
for each job to download the same pip packages, clone the same git
repos, and build the same wheels. This should help with the random
network failures that happened in the past.