Separate error checking for l0 module and l0 kernel
The macro ZE_CHECK calls the function gpuAssert which sets the Python error string if an error occurred. If no error occurs when gpuAssert is called, the Python error indicator is wiped out. Therefore we need to check the Python error indicator after create_module to ensure a gpuAssert is properly handled.

Also added an assert call to create_function to ensure a non-null module handle is provided.