[CI] Python 3.9 for PR commits, Python 3.9 - 3.11 for merge commits (#1027)
Run tests only for one Python version 3.9 for PR commits, run tests for
Python 3.9, 3.10, 3.11 for merge commits to the default branch. Since
many developers use 3.9 for their environment it is reasonable to use
3.9 as a "default" Python version for all commits. Later, after
upgrading development environments to Python 3.9 or 3.11, we can change
the "default" Python version for CI.

To merge this PR our branch protection rule that requires check for
Python 3.10 needs to be removed.