[GEN] Add `sub_group_reduce` operator (#1214)
The `gen.sub_group_reduce` operation is invoked by all work items in a
subgroup, each of them providing a $value. The $size argument is used to
form groups of $size consecutive work items called clusters. Each
cluster performs the reduction operation identified by $kind. The result
of the cluster reduction is propagated to the work items belonging to
that cluster.
It lowers to either WaveAll or WaveClustered depending on the given
size.

---------

Signed-off-by: Whitney Tsang <whitney.tsang@intel.com>