Introduce standalone runner for Triton SPIRV Kernels (#1215)
Adds skeleton for a standalone runner for SPIRV kernels generated by
Triton. The runner aims to re-use as much existing Triton code as
possible. Eventually, our goal is to link the generated `spirv_utils.so`
bundle for kernel launch code resulting in little to no code duplication
between Triton and the runner. The end state will allow a user to take a
triton cache directory containing the `_kernel.spv`, a metadata file
specifying parameters used for that kernel, and the `spirv_utils.so`
library and run the kernel under the same conditions as executed in
Triton to generate the cache. The only modification required for Triton
will be setting an environment variable to ensure the kernel parameter
metadata is placed in the cache when the cache is generated.

Currently, the runner expects input Tensors to be in Torch format. This
tracks closely with Triton and minimizes conversion issues serializing
python to c++ (or vice versa). Once XPU support for PyTorch lands in
upstream, we can switch to using Torch to move the Tensors to XPU, and
simply pass the XPU pointer to the SPIRV kernel in the runner just like
we do in Triton.