Set a default/custom skip list in CI (#1460)
Add a new workflow input `skip_list` to specify a skip list when
starting a workflow manually. If `skip_list` is not set, then workflow
checks if there is a skip list for a given `driver_version` (for example
`lts`) and selects it if it exists. Uses a default skip list in other
cases. In future, we will potentially need to add a GPU part to the skip
list name, for example `a770-lts`.

Required for #986.