Remove spirv-dis from source control (#1527)
Removing `spirv-dis` from source control (two places) since it is not
required for Triton runtime and for sure it should not be in Python
package.

We use `spirv-dis` in tests. For that, we can use `spirv-dis` from
`PATH`, this can be overridden by setting `TRITON_SPIRV_DIS_PATH`. Added
helper function to install `spirv-dis` in a development environment and
in CI workflow in case if it does not exists in `PATH`.

Fixes #647, #1530.