Create DPAS Analysis infrastructure (#1558)
Create a `DPASAnalysis` class that analyzes a function (e.g. kernel) and
determines whether `tt.dot` operations can be lowered to DPAS
instructions on ATS/PVC target. In this PR we replace uses of
`supportDPAS` with a public member function provided by the new class. A
future PR will use the same infrastructure to change the subgroup size
in the module if the analysis detects that a kernel contains `tt.dot`
operations that would be optimizable (could be lowered to DPAS
instructions) if the subgroup size was the one supported by the DPAS
instruction for the target architecture (e.g. 16 on PVC).

---------

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>