Fix the issue in the test_line_info cases. (#1782)
The root cause is the Triton doesn't support the nested layout such as
chained layout slice->dot->dpas, `#triton_gpu.slice<{dim = 1, parent =
#triton_gpu.dot_op<{opIdx = 0, parent = #mma}>}>`

It is not hit on NV and AMD backends. 
This is only possible on Intel backend when we use the Intel
RemoveLayout pass to backward propegate the DPAS layout to the ops all
the way back to the `tt.load`

To fix this: we need to implement the `getElemsPerThread` API of the
DotOp layout of DPAS layout. It allows the slice layout to calculate the
offsets per value correctly.

It should be no longer an issue after the Triton migrate to the
LinearLayout for code lowering.