[Performance] Add pass to annotate the tt.load operation with extra Intel block IO information. (#1889)
Add a pass to analysis the tt.load operations and annotate it with extra
Intel block IO information.

During the lowering phase, the block pointer type is converted to the
composed struct type. It is not easy to analysis the stride value and
pitch value to decide the matrix is column major or row major.

These information is much easier deduced at TTGIR optimization phase. To
add a new pass to annotate the `tt.load` operation with the extra
information which is required in the lowering phase but could not be
deduced easily then.

---------

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>
Co-authored-by: Tiotto, Ettore <ettore.tiotto@intel.com>