[CI][Microbenchmarks] Continue benchmarking even if one of benchmarks failed (#1982)
PR will allow benchmark run to continue even if one of the benchmarks
failed. We need complex if statement to check that installation was
successful, and that run was not cancelled.

At the same time workflow status will reflect if benchmarking was
successful, so will get red cross when something is broken.

At the end we'll have necessary artefacts. Further uploading (in
separate repo) to DB will fail to load right now because it only works
with successful runs. I'll fix it next.