Use `dest . srcLayout^-1` to preserve surjectivity for transferWithinThread conversion (#2039)
When lowering from TTGIR to LLVMIR the `convert_layout` op now uses
LinearLayout functions to compute the transformation. For a layout
conversion between elements of the same thread, i.e. a simple
reordering/permutation, we compute a `conversion` LinearLayout (LLL
function to determine the ordering, then unpack the inputs and re-pack
the outputs.

Previously, this `conversion` LL was computed using
`LL.invertAndCompose` which computes `C = A.invertAndCompose(B)` with
the property that `A(x) = B(C(x))`. So, if we compute `conversion =
src.invertAndCompose(dst)` then `src(x) = dest(conversion(x))`. But, the
`conversion` in this case is not surjective if the `src` has fewer
registers than the `dst`.

In #2004, we convert from 
```
 #blocked1 = #triton_gpu.blocked<{sizePerThread = [1, 1], threadsPerWarp = [1, 32], warpsPerCTA = [1, 4], order = [0, 1]}>
```
to
```
#blocked = #triton_gpu.blocked<{sizePerThread = [2, 1], threadsPerWarp = [1, 32], warpsPerCTA = [1, 4], order = [0, 1]}>
```

In this example, `src` has fewer registers than `dst` and so the LLVM
output values vector was only partially populated. This resulted in an
assertion and eventual segfault when NULL LLVM values were accessed
(`1`, `3`, `5`, and `,7` are missing):

```
Ins:
0 : %643 = "llvm.extractvalue"(%641) <{position = array<i64: 0>}> : (!llvm.struct<(f16, f16, f16, f16)>) -> f16
1 : %644 = "llvm.extractvalue"(%641) <{position = array<i64: 1>}> : (!llvm.struct<(f16, f16, f16, f16)>) -> f16
2 : %645 = "llvm.extractvalue"(%641) <{position = array<i64: 2>}> : (!llvm.struct<(f16, f16, f16, f16)>) -> f16
3 : %646 = "llvm.extractvalue"(%641) <{position = array<i64: 3>}> : (!llvm.struct<(f16, f16, f16, f16)>) -> f16
Outs size: 8
0 : (0, %643 = "llvm.extractvalue"(%641) <{position = array<i64: 0>}> : (!llvm.struct<(f16, f16, f16, f16)>) -> f16)
2 : (1, %644 = "llvm.extractvalue"(%641) <{position = array<i64: 1>}> : (!llvm.struct<(f16, f16, f16, f16)>) -> f16)
4 : (2, %645 = "llvm.extractvalue"(%641) <{position = array<i64: 2>}> : (!llvm.struct<(f16, f16, f16, f16)>) -> f16)
6 : (3, %646 = "llvm.extractvalue"(%641) <{position = array<i64: 3>}> : (!llvm.struct<(f16, f16, f16, f16)>) -> f16)
```

By instead inverting the source layout, we can get `conversion =
dst.invertAndCompose(src)` where `dest(x) = src(conversion(x))`. This
conversion is surjective and we can simply populate the destination by
iterating its input dimension and applying the layout to each input
value.

I also took the opportunity to make the code in
`ConvertLayoutOpToLLVM.cpp` more exactly match upstream where possible.

Resolves #2004

---------

Co-authored-by: Whitney Tsang <whitney.tsang@intel.com>