[TritonIntelGPU] Add `sub_group_transpose` operation
Add operation simulating a layout conversion such as:

For a distribution of tensor elements across a sub-group forming a
squared matrix like:

```
| Sub-group local id |                       Elements                        |
|--------------------|-------------------------------------------------------|
| 0                  | [0,sub_group_size)                                    |
| 1                  | [sub_group_size,2*sub_group_size)                     |
| ...                | ...                                                   |
| sub_group_size-1   | [(sub_group_size-1)*sub_group_size,sub_group_size**2) |
```

Redistribute elements such as the following mapping applies:

```
| Sub-group local id |                           Elements                           |
|--------------------|--------------------------------------------------------------|
| 0                  | 0, sub_group_size, 2*sub_group_size, ...                     |
| 1                  | 1, sub_group_size+1, 2*sub_group_size+1, ...                 |
| ...                | ...                                                          |
| sub_group_size-1   | [sub_group_size-1+i*sub_group_size|i in [0, sub_group_size)] |
```

In order to do this traspose, each sub-group stores its contents to
local memory via `$local_buffer`, each of them forming a matrix there,
and loads them back transposed. Note no barriers will be needed as no
memory is shared across sub-groups.

Also note the tensor shape does not change, as only the encoding of
the tensor is modified.

**NOTE:** This operation shouldn't be needed, as simply modifying the
encoding would be equivalent. As we are not handling encodings much,
we need this in the advanced path.

Signed-off-by: Victor Perez <victor.perez@codeplay.com>