Force the TRITON_XPU_GEN_NATIVE_CODE env if not previously set
This is a bit of a hack, but I did not want to pollute build_flags with something that is not a real build_flag and there is no other way to pass params to loadBinary. We could probably duplicate loadBinary - I am considering that too, but for now the hack should let the tests run.