Use stride instead of order to determine block attr (#2349)
Per the Triton slack, `order` is unused on architecture below Hopper.
But more importantly, order provides information that stride already
has. In fact, order can be completely different from stride (i.e. wrong)
and we still generate correct code. I think it is better to use the
stride assuming the logic I added here makes sense.

Note this depends on #2348, I'd like to land the debug logging
separately, so we have it even if we decide to modify this approach. It
was very useful in debugging this problem.

cc #2347