Add A^t@B benchmark (#2430)
Based on this feedback
https://github.com/intel/intel-xpu-backend-for-triton/pull/2408#pullrequestreview-2348393388

Changed GEMM benchmark to include transposed matrices case.

Closes https://github.com/intel/intel-xpu-backend-for-triton/issues/2424
Relates to
https://github.com/intel/intel-xpu-backend-for-triton/issues/1795

A@B^t case is important because weight matrix is often stored in [M, K]
format. For example, in
https://pytorch.org/docs/stable/generated/torch.nn.Linear.html
Right now we are about 1.5 times slower on XPU against raw torch for
that case.


A^t@B case is important because it's part of matmul backprop. Right now
we are about 4 times slower on XPU against raw torch for that case.