Use block load attribute to remove duplicate logic from MaterializeBlockPointer pass (#2420)
The logic in `shouldRemove` in the `RewriteTensorPointer` pass
duplicates the same logic in `MaterializeBlockPointer`:
https://github.com/intel/intel-xpu-backend-for-triton/blob/main/third_party/intel/lib/TritonIntelGPUTransforms/MaterializeBlockPointer.cpp#L50

This duplication is necessary for the Matrix transpose multiplication
case because the block pointer is defined outside a `scf.for` loop, but
the load is inside the loop. The previous logic in
`RewriteTensorPointer` could not "see" into the `scf.for` loop block and
decided to remove the tensor pointer even though its result was used by
a block load. 

This commit changes the algorithm:

First, walk the tree and look for MakeTensorPtr ops. For each MakeTensorPtr op, we do a search to find load/store users of the op. If we have a store associated with DPAS layout, or a block load, then we do not mark the MakeTensorPtr op for removal. Otherwise, we mark it for removal.

Next, we make a pass through all the ops again and make sure we removal all MakeTensorPtr-related ops for each MakeTensorPtr marked for removal (tt.advance, rewrite the loads, etc).
 
Close #2380