Use oneAPI version in PyTorch cache key
In new runners `/opt/intel/installed.txt` will contain a list of
installed packages. PyToch cache key will use this file, if exists, to
calculate a composite cache key for PyTorch. The key should be the same
for the existing runners (with PTDB), and a new PyTorch build should be
triggered on the new runners (with DLE).