[XPU][Membar] Define basic Intel-specific `Membar` filter
Define Intel-specific `Membar` filter to reduce synchronization overhead in the
Intel backend.

`Membar` analysis will insert barriers to avoid race conditions between operations
accessing the same memory. The granularity of this analysis is a bit coarse, tho,
and unneeded barriers may be inserted. In order to avoid this, the analysis
allows users to pass a callback function to filter safe cases.

As one of the recurring cases of barriers being inserted when not needed, detect
back to back layout conversions implemented as sub-group transposes as safe so no
barriers are inserted.

See code for further details on why this is safe.

Signed-off-by: victor-eds <victor.perez@codeplay.com>