Insert freeze between masked loads and sdiv/srem instructions  (#2775)
Close #2726

From the code comments:

The Triton masked load pattern can generate instances where the
mask value causes undefined behavior in sdiv/srem instructions. The
language allows this UB as the result of those arithmetic
instructions is never used, and control flow to avoid computation
of these instructions would negatively affect performance. But,
LLVM SimplifyCFG aggressively marks code paths with undefined
behavior as dead. This can result in removal of the mask path and
incorrect results from legal Triton kernels due to masked elements
being used in computation. Run a pass to add a freeze instruction
between masked loads and sdiv/srem to signal to LLVM we consider
the sdiv/srem operands to be well defined.

The strategy here is to basically invalidate the assumptions under which
SimplifyCFG can remove UB for sdiv/srem. The rationale is that, unlike
C/C++, Triton explicitly allows UB in sdiv/srem instructions (likely
because the hardware Triton is targeting allows that). Inserting a
`freeze` instruction both signals that we expect the behavior of
sdiv/srem to be well defined and hides the constant 0 in the phi from
SimplifyCFG's UB optimizations.

The pass needs to run after every instance of `InstCombine` because the
LLVM optimization that removes UB only occurs if the sdiv/srem are in
the same BB as the phi, which can happen after any `InstCombine`.

Note that the directory structure for this pass is a little different
than `BreakStructPhiNodesPass` because we are already using those
directories in `third_party` for MLIR code. If we want to change that, I
can open an issue but let's do it separately from this PR.

---------