[Layouts] Fix `invertAndCompose` when there are identity dimensions (#5468)
I found a counterexample where `invertAndCompose` was giving an
incorrect result. Thanks for @lezcano for pointing out a possible
solution: if input dim 0 and 4 are identity dimensions, then they are
added back to the reduced inverse as dimensions 3 and 4. Them reshaping
the result will assign the wrong bases to the wrong dimensions. Use a
transpose instead to order the input and output dimensions correctly
instead.