[Windows] Add `exe_extension` for `dst_path` argument of `download_and_copy` function (#3351)
https://github.com/intel/intel-xpu-backend-for-triton/actions/runs/13155450398

This fixes `RuntimeError: Cannot find ptxas-blackwell.exe`. However,
this highlights a new problem: `PermissionError: [WinError 32] The
process cannot access the file because it is being used by another
process: ''...\\AppData\\Local\\Temp\\tmpdm3gg6j5.ptx''`. The problem is
most likely the same as with `pyd` files, but it will be easier to just
skip these tests for the xpu backend.

Signed-off-by: Anatoly Myachev <anatoly.myachev@intel.com>