[triton-raise-block-ptr]: Loop Specialization (#3476)
When a `scf.for` loop contains masked loads with a mask that evaluates
to true for the entire loop iteration, the loop is versioned using a
simplified mask condition (that is loop invariant). The loads in the
loop that executes when the versioning condition evaluates to `true` are
transformed into unmasked loads and furthermore changed into loads that
operate on block pointers.

---------

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>