Loop Specialization for removal of masked `tt.load` operations (#3516)
Given an `scf.for` loop, this new pass analyzes it to determine whether
the `tt.load` operations in the loop have a mask that can be dropped by
versioning the loop.

The first cut of the pass considers `scf.for` loops that have a
'canonical' upper bound of the form `(N+END-1)/END,` where `END` is an
integer constant. Also, masked conditions must be of the form `(0..END)
< N - i*END`, where `i` is the loop induction variable.

---------

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>