[ElementwiseOpToLLVM] Do not call 'maybeDeduplicate' for dpas encoding (#3508)
Closes #3380 

The problem occurs when we try to lower an `arith.andi` op in a chain of
`tl.dot + tl.where + tl.dot`. The `maybeDeduplicate` function seems to
handle dot/dpas layouts incorrectly. When trying to deduplicate mask
values for each 16-col block it seems to mishandle strides & constancy
somehow leading to the mask values for the second 16-col block to be
ignored.

Signed-off-by: dchigarev <dmitry.chigarev@intel.com>