Enable the 2d block IO for tensor of pointer when the tensor on memory are contiguous. (#3482)
### General
This is an alternative way to lower the `tt.load` with tensor of
pointers to the 2D block IO.
It depends on the analysis result from the `ModuleAxisInfoAnalysis`
about the pointers and masks.

### Background
Both the block pointer type (e:g: `!tt.ptr<tensor<64x32xf16>`) and the
tensor of pointer type (e.g: `tensor<64x32x!tt.ptr<f16>>`) are used to
describe a tensor resident on global memory.
The difference is that the tensor of pointers contains more "entropy"
than block pointer. The tensor of pointers can describe a tensor on
global memory which is randomly distributed, of cause it can be used to
describe the structed distribution as block pointer.

There already is a optimization pass to raise the tensor of pointer to
the block pointer for some cases with limitation.
This way supports more cases with less limitation to lower the `tt.load`
with tensor of pointer to 2D block IO.

### Idea
The `ModuleAxisInfoAnalysis` analysis the axis information such as,
contiguity, divisibility and constancy of the value of the tensor for
pointers and masks.
In general, we can use the 2D block IO lowering as long as:
1. the contiguity of the pointers is multiple of the threadsPerWarp size
in the same Dim D.
2. The constancy of the masks is multiple of the threadsPerWarp size in
the same Dim D.

We start with the case of the `tt.load` with DotOp and DPAS layout. To
generalize the lowering code for more cases with LinearLayout in future.