[Windows] Fix `level_zero/ze_api.h` not found when running UT by using `LEVEL_ZERO_V1_SDK_PATH` env var (#3656)
@pbchekin @vlad-penkin I took all the artifacts after using (on local
machine) the sdk installer (all options by default), packed the folder
into a tar archive and changed the name of the environment variable. It
seems that this is all that is needed for integration.

**UPD**: although `ZE_PATH` is not a standard and documented variable in
level_zero (at least I haven't found it), it was decided to leave it in
so as not to break the code of users who might use it. Interestingly, I
haven't found any use of it in the PyTorch code.

---------

Signed-off-by: Anatoly Myachev <anatoly.myachev@intel.com>
(cherry picked from commit 9304351d9c4a51cdb23f113415ef55a149126a74)