Remove redundant spirv attribute specifying the 'subgroup_size' (#3635)
# Description

This PR removes the SPIR-V env containing the `subgroup_size` attribute
that was injected during the lowering from TritonIntelGPU to LLVM,
originally introduced in [PR
#1270](https://github.com/intel/intel-xpu-backend-for-triton/pull/1270).
The removal addresses the concern raised in [Issue
#1313](https://github.com/intel/intel-xpu-backend-for-triton/issues/1313).

## Background

The `subgroup_size` attribute was initially required by the
`SubGroupReduce`'s verifier, as discussed in [Issue
#1235](https://github.com/intel/intel-xpu-backend-for-triton/issues/1235),
and was introduced in a series of PRs:

- [PR
#1223](https://github.com/intel/intel-xpu-backend-for-triton/pull/1223)
- [PR
#1262](https://github.com/intel/intel-xpu-backend-for-triton/pull/1262)
- [PR
#1270](https://github.com/intel/intel-xpu-backend-for-triton/pull/1270)

In PR #1223, the `TritonGen::getSubgroupSize` function was introduced to
retrieve the subgroup size from the `spirv.target_env`. However,
questions arose regarding whether this attribute was redundant, as
`ttg.threads-per-warp` already provided the necessary information.

A first attempt to remove this duplication was made in [PR
#1241](https://github.com/intel/intel-xpu-backend-for-triton/pull/1241),
but it was ultimately closed without merging.

## Why Can We Remove It Now?

1. Redundant Getter Function Replaced:
- A more recent [PR
#3512](https://github.com/intel/intel-xpu-backend-for-triton/pull/3512)
replaced `TritonGen::getSubgroupSize` to use `ttg.threads-per-warp`
instead of the `spirv.target_env` attribute.
2. Verifier No Longer Requires It:
- The last remaining use of the SPIR-V `subgroup_size` attribute was in
the verifier for the `SubGroupReduce` op.
- This became obsolete when [PR
#2893](https://github.com/intel/intel-xpu-backend-for-triton/pull/2893)
removed the use of TritonGen dialect's `subgroup_reduce` op, replacing
it with an equivalent SPIR-V op.

## Conclusion

With no remaining usages of the SPIR-V `subgroup_size` attribute, it is
now safe to remove it entirely, simplifying the IR and avoiding
unnecessary redundancy.

<!---
The core Triton is a small number of people, and we receive many PRs
(thank
you!).  To help us review your code more quickly, **if you are a new
contributor (less than 3 PRs merged) we ask that you complete the
following
tasks and include the filled-out checklist in your PR description.**

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them.
-->

# New contributor declaration
- [x] I am not making a trivial change, such as fixing a typo in a
comment.

- [x] I have written a PR description following these
  [rules](https://cbea.ms/git-commit/#why-not-how).

- [x] I have run `pre-commit run --from-ref origin/main --to-ref HEAD`.

- Select one of the following.
  - [ ] I have added tests.
    - `/test` for `lit` tests
    - `/unittest` for C++ tests
    - `/python/test` for end-to-end tests
- [x] This PR does not need a test because `it only removes an unused
attribute`.

- Select one of the following.
  - [x] I have not added any `lit` tests.
- [ ] The `lit` tests I have added follow these [best
practices](https://mlir.llvm.org/getting_started/TestingGuide/#filecheck-best-practices),
including the "tests should be minimal" section. (Usually running Python
code
    and using the instructions it generates is not minimal.)