Fix OptimizeReductionLocality accuracy issue
Reorder the pass to avoid the RemoveLayoutConversion pass to remove needed LayoutConvertOp.
Re-introduce a condition to ReshapeOp canonicalizer to avoir the explicitely needed LayoutConvertOp to be removed.
Add a extra condition to the OptimizeReductionLocality pass to avoid reshape problem if number of elements smaller than the encoding.
Add a test to check the new condition in the pass.
Enhance the flash-attention helper script to check and validate the outputs.