[tensor_desc_to_block_ptr]: Translate `tt.descritor_store` into `tt.store` (#3778)
This is a continuation of the work started in PR #3756, specifically
this PR adds support for translation of a `tt.descriptor_store`
operation into a `tt.store` operation. The associated
`tt.make_tensor_descriptor` operation is replaced with a
`tt.make_block_ptr` operation (where the offsets argument are inferred
from the `tt.descritor_store` operation).

---------

Signed-off-by: Tiotto, Ettore <ettore.tiotto@intel.com>