Add functional test for 2D block load lowering (#3780)
Adds a functional test for 2d block load lowering. The lit tests do not
catch changes to the arguments of the 2D block load (e.g. indexing)
because they do not check for every single generated instruction. The
functional test for lowering works by generating TTGIR which will be
lowered to 2D block load instructions, then using a 2D block store to
store the result and compare with the input. The test supports A, B
(vnni transform), and B transpose loads.

close #3764