[CI][benchmarks] Add `n_runs` parameter (#3800)
Closes https://github.com/intel/intel-xpu-backend-for-triton/issues/3163

Verification only on the first run is not implemented yet, I suggest to
do that in separate PR to limit delta size. It will require further
redesign of Benchmark class.

Further suggestions:
1. Implement verification flag as an argument for all of our benchmarks
and as a CLI argument. Use it for `n_runs > 1`. Use it in workflow for
verification functionality
2. Make `build_report` script into a function that each benchmark calls
by itself to remove it from the workflow